Add int casting

+++ median.py
@@ -14,8 +14,8 @@ numbers.sort()
 length = len(numbers)
 
 if length % 2 == 1:
-    print(numbers[(length-1) / 2])
+    print(numbers[int((length-1) / 2)])
 else:
-    median = ((numbers[length / 2]) + (numbers[length / 2 - 1])) / 2
+    median = (numbers[int(length / 2)] + numbers[int(length / 2 - 1)]) / 2
     print(median)
